requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
altair>=5.0.0
Pillow>=10.0.0
